Final delivery code of Room Management System (RMS)
Finding a proper place for your team to work on a project or any
discussion is very important. Perhaps, the most common place found is at
the library which contains limited space comparing to a large body of
students. Also, the conventional method to get a space there is to book
through the help of librarian officers which adding more work to both
the students and the officers in term of locating free room and at a
convenient time for their meeting.
This is why we come up with a project called “Room Management System”.
It will solve these problems with its main functionalities such as:
- Self-locating room, not only at the library but also in other
classroom building.
- Monitoring classroom used by the students more effectively and
efficiently with one click as Approve or Reject.
- Keeping track of how actively students work together as a team or
individually throughout term, semester or academic year.
These will result in better understanding of future focusing development
and research of the university in many parts through the report
generating from the data over years such as the demands of student
working space, the current trend topics of discussions, and more.
In Conclusion, the project will play important role in making the most
use of available space on campus for educational purposes among students
and faculty.